Add form handling for inserting and editing data entities
Introduced various forms to handle data insertion across multiple components, including "Utenze," "Edificio," and "Fotovoltaico." Implemented validation using `zod` and updated database DAOs to support the new attributes.

+++ src-tauri/src/dto/fotovoltaico_dto.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 pub struct FotovoltaicoDTO {
     pub(crate) id: Option<u64>,
     pub(crate) id_edificio: String,
-    pub(crate) potenza: u16,
+    pub(crate) potenza: f32,
     pub(crate) proprietario: String,
 }
